pubsub/requirements.txt
Flask==0.12.2
google-cloud-pubsub==0.33.0
gunicorn==19.7.1
requests>=2.18.0
pip==18.0
